Remove person module and update customer schema
