--- customer-api/e2e/delete-person.http
@@ -1,9 +1,0 @@
-# App: Customer CRUD Application
-# Package: customer-api
-# File: e2e/delete-person.http
-# Version: 2.0.34
-# Author: Bobwares
-# Date: 2025-06-05 00:14:14 UTC
-# Description: E2E test for deleting a person.
-#
-DELETE http://localhost:3001/persons/1
