--- customer-api/e2e/update-person.http
@@ -1,14 +1,0 @@
-# App: Customer CRUD Application
-# Package: customer-api
-# File: e2e/update-person.http
-# Version: 2.0.34
-# Author: Bobwares
-# Date: 2025-06-05 00:14:14 UTC
-# Description: E2E test for updating a person.
-#
-PATCH http://localhost:3001/persons/1
-Content-Type: application/json
-
-{
-  "age": 31
-}
