+++ customer-api/src/app.module.ts
@@ -1,9 +1,9 @@
 // App: Customer CRUD Application
 // Package: customer-api
 // File: src/app.module.ts
-// Version: 2.0.34
+// Version: 2.0.35
 // Author: Bobwares
-// Date: 2025-06-05 00:13:21 UTC
+// Date: 2025-06-05 01:22:19 UTC
 // Description: Root application module.
 //
 import { Module } from '@nestjs/common';
@@ -11,8 +11,6 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule } from '@nestjs/config';
 import { CustomersModule } from './customers/customers.module';
 import { Customer } from './customers/customer.entity';
-import { PersonsModule } from './persons/persons.module';
-import { Person } from './persons/person.entity';
 
 @Module({
   imports: [
@@ -24,11 +22,10 @@ import { Person } from './persons/person.entity';
       username: process.env.DB_USER,
       password: process.env.DB_PASS,
       database: process.env.DB_NAME,
-      entities: [Customer, Person],
+      entities: [Customer],
       synchronize: true,
     }),
     CustomersModule,
-    PersonsModule,
   ],
 })
 export class AppModule {}
