+++ customer-api/src/customers/customers.service.ts
@@ -1,9 +1,9 @@
 // App: Customer CRUD Application
 // Package: customer-api
 // File: src/customers/customers.service.ts
-// Version: 2.0.34
+// Version: 2.0.35
 // Author: Bobwares
-// Date: 2025-06-05 00:14:14 UTC
+// Date: 2025-06-05 01:23:52 UTC
 // Description: Business logic for managing customers.
 //
 import { Injectable, NotFoundException } from '@nestjs/common';
@@ -18,8 +18,21 @@ export class CustomersService {
     private readonly repo: Repository<Customer>,
   ) {}
 
-  async create(data: { name: string; email: string }): Promise<Customer> {
-    const customer = this.repo.create(data);
+  async create(data: {
+    first: string;
+    last: string;
+    age: number;
+    address: { street: string; city: string; state: string; zipcode: string };
+  }): Promise<Customer> {
+    const customer = this.repo.create({
+      first: data.first,
+      last: data.last,
+      age: data.age,
+      street: data.address.street,
+      city: data.address.city,
+      state: data.address.state,
+      zipcode: data.address.zipcode,
+    });
     return this.repo.save(customer);
   }
 
@@ -35,13 +48,43 @@ export class CustomersService {
     return customer;
   }
 
-  async update(id: number, data: { name?: string; email?: string }): Promise<Customer> {
+  async update(
+    id: number,
+    data: {
+      first?: string;
+      last?: string;
+      age?: number;
+      address?: {
+        street?: string;
+        city?: string;
+        state?: string;
+        zipcode?: string;
+      };
+    },
+  ): Promise<Customer> {
     const customer = await this.findOne(id);
-    if (data.name !== undefined) {
-      customer.name = data.name;
+    if (data.first !== undefined) {
+      customer.first = data.first;
     }
-    if (data.email !== undefined) {
-      customer.email = data.email;
+    if (data.last !== undefined) {
+      customer.last = data.last;
+    }
+    if (data.age !== undefined) {
+      customer.age = data.age;
+    }
+    if (data.address) {
+      if (data.address.street !== undefined) {
+        customer.street = data.address.street;
+      }
+      if (data.address.city !== undefined) {
+        customer.city = data.address.city;
+      }
+      if (data.address.state !== undefined) {
+        customer.state = data.address.state;
+      }
+      if (data.address.zipcode !== undefined) {
+        customer.zipcode = data.address.zipcode;
+      }
     }
     return this.repo.save(customer);
   }
