+++ customer-api/src/customers/dto/update-customer.dto.ts
@@ -1,9 +1,9 @@
 // App: Customer CRUD Application
 // Package: customer-api
 // File: src/customers/dto/update-customer.dto.ts
-// Version: 2.0.34
+// Version: 2.0.35
 // Author: Bobwares
-// Date: 2025-06-05 00:14:14 UTC
+// Date: 2025-06-05 01:23:26 UTC
 // Description: DTO for updating a customer.
 //
 import { PartialType } from '@nestjs/mapped-types';
