--- customer-api/src/persons/dto/update-person.dto.ts
@@ -1,12 +1,0 @@
-// App: Customer CRUD Application
-// Package: customer-api
-// File: src/persons/dto/update-person.dto.ts
-// Version: 2.0.34
-// Author: Bobwares
-// Date: 2025-06-05 00:13:21 UTC
-// Description: DTO for updating a person.
-//
-import { PartialType } from '@nestjs/mapped-types';
-import { CreatePersonDto } from './create-person.dto';
-
-export class UpdatePersonDto extends PartialType(CreatePersonDto) {}
