+++ customer-app/src/app/customers/page.tsx
@@ -1,27 +1,41 @@
 // App: Customer CRUD Application
 // Package: customer-app
 // File: src/app/customers/page.tsx
-// Version: 2.0.34
+// Version: 2.0.35
 // Author: Bobwares
-// Date: 2025-06-05 00:14:14 UTC
+// Date: 2025-06-05 01:26:26 UTC
 // Description: Customer maintenance page using customer-api backend.
 //
 "use client";
 import { useEffect, useState } from 'react';
 import styles from './page.module.css';
 
+interface Address {
+  street: string;
+  city: string;
+  state: string;
+  zipcode: string;
+}
+
 interface Customer {
   id: number;
-  name: string;
-  email: string;
+  first: string;
+  last: string;
+  age: number;
+  address: Address;
 }
 
 const API_URL = 'http://localhost:3001/customers';
 
 export default function CustomersPage() {
   const [customers, setCustomers] = useState<Customer[]>([]);
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
+  const [form, setForm] = useState<Customer>({
+    id: 0,
+    first: '',
+    last: '',
+    age: 0,
+    address: { street: '', city: '', state: '', zipcode: '' },
+  });
   const [editingId, setEditingId] = useState<number | null>(null);
 
   const loadCustomers = async () => {
@@ -35,24 +49,30 @@ export default function CustomersPage() {
   }, []);
 
   const resetForm = () => {
-    setName('');
-    setEmail('');
+    setForm({
+      id: 0,
+      first: '',
+      last: '',
+      age: 0,
+      address: { street: '', city: '', state: '', zipcode: '' },
+    });
     setEditingId(null);
   };
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
+    const payload = { ...form };
     if (editingId === null) {
       await fetch(API_URL, {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ name, email }),
+        body: JSON.stringify(payload),
       });
     } else {
       await fetch(`${API_URL}/${editingId}`, {
         method: 'PATCH',
         headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ name, email }),
+        body: JSON.stringify(payload),
       });
     }
     resetForm();
@@ -62,8 +82,7 @@ export default function CustomersPage() {
   const handleEdit = (id: number) => {
     const customer = customers.find((c) => c.id === id);
     if (!customer) return;
-    setName(customer.name);
-    setEmail(customer.email);
+    setForm(customer);
     setEditingId(id);
   };
 
@@ -75,17 +94,75 @@ export default function CustomersPage() {
     loadCustomers();
   };
 
+  const setAddress = (key: keyof Address, value: string) => {
+    setForm((prev) => ({
+      ...prev,
+      address: { ...prev.address, [key]: value },
+    }));
+  };
+
   return (
     <div className={styles.container}>
       <h1>Customer Maintenance</h1>
       <form onSubmit={handleSubmit} className={styles.form}>
         <label>
-          Name
-          <input value={name} onChange={(e) => setName(e.target.value)} required />
+          First
+          <input
+            value={form.first}
+            onChange={(e) => setForm({ ...form, first: e.target.value })}
+            required
+          />
         </label>
         <label>
-          Email
-          <input value={email} onChange={(e) => setEmail(e.target.value)} required />
+          Last
+          <input
+            value={form.last}
+            onChange={(e) => setForm({ ...form, last: e.target.value })}
+            required
+          />
+        </label>
+        <label>
+          Age
+          <input
+            type="number"
+            value={form.age}
+            onChange={(e) => setForm({ ...form, age: Number(e.target.value) })}
+            min={0}
+            required
+          />
+        </label>
+        <label>
+          Street
+          <input
+            value={form.address.street}
+            onChange={(e) => setAddress('street', e.target.value)}
+            required
+          />
+        </label>
+        <label>
+          City
+          <input
+            value={form.address.city}
+            onChange={(e) => setAddress('city', e.target.value)}
+            required
+          />
+        </label>
+        <label>
+          State
+          <input
+            value={form.address.state}
+            onChange={(e) => setAddress('state', e.target.value)}
+            required
+            maxLength={2}
+          />
+        </label>
+        <label>
+          Zipcode
+          <input
+            value={form.address.zipcode}
+            onChange={(e) => setAddress('zipcode', e.target.value)}
+            required
+          />
         </label>
         <div className={styles.formButtons}>
           <button type="submit" className={styles.button}>
@@ -105,16 +182,26 @@ export default function CustomersPage() {
         <table className={styles.table}>
           <thead>
             <tr>
-              <th>Name</th>
-              <th>Email</th>
+              <th>First</th>
+              <th>Last</th>
+              <th>Age</th>
+              <th>Street</th>
+              <th>City</th>
+              <th>State</th>
+              <th>Zipcode</th>
               <th>Actions</th>
             </tr>
           </thead>
           <tbody>
             {customers.map((c) => (
               <tr key={c.id}>
-                <td>{c.name}</td>
-                <td>{c.email}</td>
+                <td>{c.first}</td>
+                <td>{c.last}</td>
+                <td>{c.age}</td>
+                <td>{c.address.street}</td>
+                <td>{c.address.city}</td>
+                <td>{c.address.state}</td>
+                <td>{c.address.zipcode}</td>
                 <td>
                   <button onClick={() => handleEdit(c.id)}>Edit</button>
                   <button onClick={() => handleDelete(c.id)}>Delete</button>
