--- customer-app/src/app/persons/page.module.css
@@ -1,72 +1,0 @@
-.container {
-  background-color: #1e3a8a; /* dark blue */
-  min-height: 100vh;
-  padding: 2rem;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.form {
-  background-color: #ffffff;
-  border: 1px solid #ccc;
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  width: 300px;
-}
-
-.form label {
-  display: flex;
-  flex-direction: column;
-  font-weight: 600;
-  gap: 0.25rem;
-}
-
-.form input {
-  padding: 0.25rem;
-}
-
-.formButtons {
-  display: flex;
-  gap: 0.5rem;
-}
-
-.button {
-  background-color: #1e3a8a;
-  color: #fff;
-  border: none;
-  padding: 0.5rem 1rem;
-  cursor: pointer;
-}
-
-.button:hover {
-  background-color: #162d6a;
-}
-
-.cancelButton {
-  background-color: #f0f0f0;
-  color: #333;
-}
-
-.cancelButton:hover {
-  background-color: #e0e0e0;
-}
-
-.tableContainer {
-  background-color: #ffffff;
-  border: 1px solid #ccc;
-  padding: 1rem;
-}
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.table th,
-.table td {
-  border: 1px solid #ccc;
-  padding: 0.5rem;
-}
